Fix V3041, V3105 warnings from PVS-Studio Static Analyzer

+++ TwitchLib/Models/Client/ChatMessage.cs
@@ -271,21 +271,21 @@
             */
             if (bits < 1500)
             {
-                return (bits / 100) * 1.4;
+                return ((double)(bits) / 100) * 1.4;
             }
             if (bits < 5000)
             {
-                return (bits / 1500) * 19.95;
+                return ((double)(bits) / 1500) * 19.95;
             }
             if (bits < 10000)
             {
-                return (bits / 5000) * 64.40;
+                return ((double)(bits) / 5000) * 64.40;
             }
             if (bits < 25000)
             {
-                return (bits / 10000) * 126;
+                return ((double)(bits) / 10000) * 126;
             }
-            return (bits / 25000) * 308;
+            return ((double)(bits) / 25000) * 308;
         }
     }
 }
